mp 0.3.0: f64 sweep + binary bincode wire (sim + WASM bindings)
Foundation for Phase 2 wire integration. All mp1 scalars are now f64
(was f32) to match JavaScript Number precision — eliminates implicit
narrowing at the wasm-bindgen boundary so WASM-side prediction stays
byte-identical to native server math.

Wire format switches from JSON-tagged to bincode binary, using the
same DefaultOptions + fixint + LE convention as the legacy protocol
codec. ~6x smaller than JSON would have been; 28-byte Input frames,
228-byte snapshots for 4 ships.

Externally-tagged enum encoding (dropped `#[serde(tag = "kind")]`)
because bincode 1.x can't deserialize internally-tagged variants
(needs `deserialize_any`, non-self-describing). Externally-tagged
works in both bincode AND JSON cleanly.

New `Snapshot.acked_input_tick` field — Phase 4 reconciliation will
replay pending inputs from this point forward when local prediction
diverges from server truth.

Files:
  server/sim/src/mp1/state.rs       — ShipState fields + field consts to f64
  server/sim/src/mp1/input.rs       — PlayerInput + tuning constants to f64
  server/sim/src/mp1/ship.rs        — f64 math (atan2/hypot/cos/sin/abs)
  server/sim/src/mp1/wire.rs        — SnapshotShip + Input + Welcome f64;
                                       removed `tag="kind"`; added
                                       acked_input_tick
  server/sim/src/mp1/codec.rs       — NEW: bincode encode/decode helpers
  server/sim/src/mp1/mod.rs         — re-export codec helpers
  server/sim/Cargo.toml             — serde_json dev-dep for JSON tests
  server/client-wasm/src/lib.rs     — World accessors + set_input f64

Tests: cargo test -p rainboids-sim — 11 pass (6 wire JSON + 5 codec
bincode + 1 cross-format invariant). cargo test --workspace — 54 pass
(no rainboids-server regressions). npm run wasm:build:dev — succeeds
(~135 KB unoptimized).

See docs/Multiplayer WASM Pivot Phase 2 - 2026-05-17.md.

+++ server/client-wasm/src/lib.rs
@@ -67,8 +67,8 @@ impl World {
         down: bool,
         left: bool,
         right: bool,
-        aim_x: f32,
-        aim_y: f32,
+        aim_x: f64,
+        aim_y: f64,
     ) {
         self.input.up = up;
         self.input.down = down;
@@ -81,41 +81,41 @@ impl World {
     /// Advance simulation by one tick. `dt` is seconds (unused by
     /// Phase-1 ship physics, which works in per-tick deltas; accepted
     /// for forward compatibility with future delta-time integration).
-    pub fn tick(&mut self, dt: f32) {
+    pub fn tick(&mut self, dt: f64) {
         tick_phase1(&mut self.state, &self.input, dt);
     }
 
     // ── Read-only accessors ──
 
-    pub fn ship_x(&self) -> f32 {
+    pub fn ship_x(&self) -> f64 {
         self.state.ship.x
     }
 
-    pub fn ship_y(&self) -> f32 {
+    pub fn ship_y(&self) -> f64 {
         self.state.ship.y
     }
 
-    pub fn ship_vx(&self) -> f32 {
+    pub fn ship_vx(&self) -> f64 {
         self.state.ship.vx
     }
 
-    pub fn ship_vy(&self) -> f32 {
+    pub fn ship_vy(&self) -> f64 {
         self.state.ship.vy
     }
 
-    pub fn ship_angle(&self) -> f32 {
+    pub fn ship_angle(&self) -> f64 {
         self.state.ship.angle
     }
 
-    pub fn ship_radius(&self) -> f32 {
+    pub fn ship_radius(&self) -> f64 {
         self.state.ship.radius
     }
 
-    pub fn field_width(&self) -> f32 {
+    pub fn field_width(&self) -> f64 {
         self.state.field_w
     }
 
-    pub fn field_height(&self) -> f32 {
+    pub fn field_height(&self) -> f64 {
         self.state.field_h
     }
