+++ server/sim/src/mp1/state.rs
@@ -2,27 +2,35 @@
 //! collections (enemies, asteroids, bullets, drops) are deferred to
 //! Phase 3+. Authored fresh from `js/modules/core/constants.js` and
 //! `js/sim/ship.js` as of 2026-05-17.
+//!
+//! All scalars are **f64** to match JavaScript Number precision. The
+//! WASM client and the native server both run the same Rust code via
+//! `mp1::tick_phase1`; f64 throughout means no implicit narrowing at
+//! the wasm-bindgen boundary. Wire-format snapshots also use f64
+//! (8 bytes/scalar vs 4 for f32) — costs ~70% more bandwidth but
+//! eliminates a class of cross-runtime drift bugs for prediction
+//! reconciliation in Phase 4+.
 
 use serde::{Deserialize, Serialize};
 
 /// World bounds — Rainboids' logical field. Pulled from
 /// `GAME_CONFIG.FIELD_WIDTH/FIELD_HEIGHT` in solo's constants.js.
-pub const FIELD_WIDTH: f32 = 1920.0;
-pub const FIELD_HEIGHT: f32 = 1080.0;
-pub const SHIP_SIZE: f32 = 30.0;
+pub const FIELD_WIDTH: f64 = 1920.0;
+pub const FIELD_HEIGHT: f64 = 1080.0;
+pub const SHIP_SIZE: f64 = 30.0;
 
 /// One ship's physics + identity state. Mirrors the fields the solo
 /// ship.js step function reads/writes (x, y, vx, vy, angle, hp).
 #[derive(Debug, Clone, Copy, Serialize, Deserialize)]
 pub struct ShipState {
-    pub x: f32,
-    pub y: f32,
-    pub vx: f32,
-    pub vy: f32,
-    pub angle: f32, // radians; 0 = +x axis (right)
-    pub hp: f32,
-    pub max_hp: f32,
-    pub radius: f32,
+    pub x: f64,
+    pub y: f64,
+    pub vx: f64,
+    pub vy: f64,
+    pub angle: f64, // radians; 0 = +x axis (right)
+    pub hp: f64,
+    pub max_hp: f64,
+    pub radius: f64,
     pub active: bool,
 }
 
@@ -35,7 +43,7 @@ impl Default for ShipState {
             y: FIELD_HEIGHT * 0.5,
             vx: 0.0,
             vy: 0.0,
-            angle: -std::f32::consts::FRAC_PI_2,
+            angle: -std::f64::consts::FRAC_PI_2,
             hp: 100.0,
             max_hp: 100.0,
             radius: SHIP_SIZE * 0.5,
@@ -48,8 +56,8 @@ impl Default for ShipState {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct GameState {
     pub ship: ShipState,
-    pub field_w: f32,
-    pub field_h: f32,
+    pub field_w: f64,
+    pub field_h: f64,
     pub tick: u32,
 }
